requirements.txt
pyTelegramBotAPI
google-genai
Pillow